third_party/WebKit/LayoutTests/geolocation-api/timeout-negative.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../resources/js-test.js"></script>
<script src="../resources/mojo-helpers.js"></script>
<script src="resources/geolocation-mock.js"></script>
</head>
<body>
<script>
description("Tests that when timeout is negative (and maximumAge is too), the error callback is called immediately with code TIMEOUT.");

var error;

geolocationMock.then(mock => {
    mock.setGeolocationPosition(51.478, -0.166, 100.0);

    navigator.geolocation.getCurrentPosition(function(p) {
        testFailed('Success callback invoked unexpectedly');
        finishJSTest();
    }, function(e) {
        error = e;
        shouldBe('error.code', 'error.TIMEOUT');
        shouldBe('error.message', '"Timeout expired"');
        finishJSTest();
    }, {
        timeout: -1000
    });
});

window.jsTestIsAsync = true;

</script>
</body>
</html>
